Rename api file
