+++ src/routes/+page.server.ts
@@ -1,5 +1,5 @@
 import { LEADERBOARD_UPDATE_INTERVAL } from '$lib/constants/data';
-import { GetLeaderboardSlice } from '$lib/eliteapi/eliteapi';
+import { GetLeaderboardSlice } from '$lib/api/elite';
 import type { PageServerLoad } from './$types';
 
 export const load: PageServerLoad = async ({ setHeaders }) => {
